Added automatic controller discovery by using reflection and mapping all public methods into the route table by convention.
Added HttpGet and HttpPost attributes to automatically register the HTTP method of the action.

Added Authorize attribute. Implemented a layout page logic and insert the view content in it.

Used reflection to analyze the action parameters and populate them automatically from the request.

Removed the required constructor on the base Controller class.